Transfer startFirestoreSync to buklod-tao - Possible future for resetting the replication sync
